Add support for running commands inside the temporary container
This commits adds two types of sub-commands:

- The "run" subcommand
- Subcommands named after the available docker images we have, e.g.
  "ubuntu_base"

Both types of subcommands takes the same number of arguments and it
allows us to launch any other arguments that follow in the command
line as a command inside a container. The docker_wrapper library
will launch the container, run the command and terminate the container
at thte end.

This commit fixes the specifying the entrypoint scripts in the
Dockerfile.