Provide a getter/setter interface API so that the environment is set later
This commit adds a set/getter interface to the configuration
state that become a global/static variable. We did that so that
the configuration can be initialized after we create the CLI
"asynchronously". This is necessary as the name of the configuration
to use will be comming from a CLI argument define by the parent
typer.