Add example entrypoing repo-cli.py script
This commit adds the repo-cli.py script that serves as an example
how docker-wrapper is to be used in a repo to provide a CLI
wrapper for docker.

Add also an example environment-cfg.yml file that could be used
for the configuration of the CLI and pass options/parameters
in downstream tools.

Fix the existing sample-images to build and work given the
changes to the CLI.