Pass **kwargs to the DockerImage constructor and in the derived class constructors
This commit removes passing the explicit docker_registry_prefix argument
through out the DockerImage and its derived class constructors.

Instead pass `**kwargs` throug the whole call chain and allow each
constructor to specify its own named arguments.

Even though this is not a good coding practice it will allow us
for how to relax the requirements of the values defined in the
`*-environment-config.yaml` file to be specific to the needs
of each docker image.