Simplify the image entrypoint, use the `-u` option
This commit simplifies the entrypoint scripts of the docker
images.
Instead of relying on the entrypoint to create a user
with a matching UID/GID as on the host do the following:

1. Mount the host's password, users, files inside the container
This way the UID/GID will exist inside the container and be the
same as on the host.
2. Use the `-u GID:PID` option to set the user inside the container
to match the user instatiating the container.
3. Add argument called "enable_sudo" ( that is on by default in
the CLI ) to also mount the /etc/sudoers.d folder.