feat(cli): Add option to enable/disable mounting the host passwd file
In this PR we are introdcing the --[no]mount_host_passwd option
that enables/diasbles mounting any password/user/group related
files from the host inside the container and making use of the
`--user` option to specify the user. This is necessary in order
 to handle cases/systems where the above files are not comprehensive
(e.g. when there is use of NIS/LDAP for user password management)